moving crypted buffer into core
