+++ src/wf_core/message.rs
@@ -1,8 +1,10 @@
-use super::decoder::Decoder;
-use super::segment::MessageSegment;
-use super::FieldValue;
+use super::{
+    crypted_buffer::{CryptMode, CryptedBuffer},
+    segment::MessageSegment,
+    Decoder, FieldValue,
+};
 use crate::wf_account::test_impl::WhiteflagAccount;
-use crate::wf_buffer::{CryptMode, CryptedBuffer, WhiteflagBuffer};
+use crate::wf_buffer::WhiteflagBuffer;
 use crate::wf_parser::{builder_from_field_values, builder_from_serialized};
 use fennel_lib::FennelCipher;
 use wf_crypto::encryption_method::WhiteflagEncryptionMethod;
